Revise the test suite of ClientServiceBase class
